Seperated Redis and data classes
