+++ pom.xml
@@ -47,7 +47,7 @@
 		<dependency>
 			<groupId>redis.clients</groupId>
 			<artifactId>jedis</artifactId>
-			<version>4.1.0</version>
+			<version>4.2.3</version>
 			<type>jar</type>
 			<scope>compile</scope>
 		</dependency>
