+++ src/main/java/de/simonsator/partyandfriends/spigot/api/party/PartyManager.java
@@ -3,14 +3,13 @@ package de.simonsator.partyandfriends.spigot.api.party;
 import de.simonsator.partyandfriends.spigot.api.pafplayers.PAFPlayer;
 import de.simonsator.partyandfriends.spigot.utilities.disable.Deactivated;
 import org.bukkit.plugin.java.JavaPlugin;
-import redis.clients.jedis.Jedis;
 
 /**
  * Manages the parties. From here you can get a {@link de.simonsator.partyandfriends.spigot.api.party.PlayerParty PlayerParty}.
  *
  * @author Simonsator
  */
-public class PartyManager extends JavaPlugin implements Deactivated {
+public abstract class PartyManager implements Deactivated {
 	private static PartyManager instance;
 
 	/**
@@ -18,12 +17,6 @@ public class PartyManager extends JavaPlugin implements Deactivated {
 	 */
 	public PartyManager() {
 		instance = this;
-	}
-
-	@Override
-	public void onEnable() {
-		getConfig().options().copyDefaults(true);
-		saveConfig();
 	}
 
 	/**
@@ -35,13 +28,6 @@ public class PartyManager extends JavaPlugin implements Deactivated {
 		return instance;
 	}
 
-	Jedis getConnection() {
-		Jedis jedis = new Jedis(getConfig().getString("Redis.Host"), getConfig().getInt("Redis.Port"));
-		if (!getConfig().getString("Redis.Password").equals(""))
-			jedis.auth(getConfig().getString("Redis.Password"));
-		return jedis;
-	}
-
 	/**
 	 * Returns null if the player is not in a party. If the player is in a party (either as the leader or as a party member),
 	 * then this function returns a {@link de.simonsator.partyandfriends.spigot.api.party.PlayerParty PlayerParty} representing the party.
@@ -50,12 +36,5 @@ public class PartyManager extends JavaPlugin implements Deactivated {
 	 * @return Returns null if the player is not in a party. If the player is in a party (either as the leader or as a party member),
 	 * then this function returns a {@link de.simonsator.partyandfriends.spigot.api.party.PlayerParty PlayerParty} representing the party.
 	 */
-	public PlayerParty getParty(PAFPlayer pPlayer) {
-		try (Jedis jedis = getConnection()) {
-			String id = jedis.hget("paf:parties:players:id", pPlayer.getUniqueId().toString());
-			if (id == null)
-				return null;
-			return new PlayerParty(Integer.parseInt(id));
-		}
-	}
+	public abstract PlayerParty getParty(PAFPlayer pPlayer);
 }
